Move some types to new packages.
