+++ platforms/core-runtime/client-services/src/main/java/org/gradle/internal/daemon/client/clientinput/DaemonClientInputForwarder.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradle.launcher.daemon.client;
+package org.gradle.internal.daemon.client.clientinput;
 
 import org.gradle.internal.UncheckedException;
 import org.gradle.internal.concurrent.Stoppable;
