+++ platforms/core-runtime/client-services/src/test/groovy/org/gradle/internal/daemon/client/serialization/ClientSidePayloadClassLoaderFactoryTest.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 the original author or authors.
+ * Copyright 2024 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package org.gradle.tooling.internal.provider.serialization
+package org.gradle.internal.daemon.client.serialization
 
 import org.gradle.internal.classloader.VisitableURLClassLoader
+import org.gradle.tooling.internal.provider.serialization.PayloadClassLoaderFactory
 import spock.lang.Specification
 
 class ClientSidePayloadClassLoaderFactoryTest extends Specification {
