+++ platforms/core-runtime/daemon-protocol/src/main/java/org/gradle/tooling/internal/provider/serialization/ClassLoaderDetails.java
@@ -25,9 +25,9 @@ import java.util.UUID;
 
 public class ClassLoaderDetails implements Serializable {
     // TODO:ADAM - using a UUID means we create a ClassLoader hierarchy for each daemon process we talk to. Instead, use the spec to decide whether to reuse a ClassLoader
-    final UUID uuid;
-    final ClassLoaderSpec spec;
-    final List<ClassLoaderDetails> parents = new ArrayList<ClassLoaderDetails>();
+    public final UUID uuid;
+    public final ClassLoaderSpec spec;
+    public final List<ClassLoaderDetails> parents = new ArrayList<ClassLoaderDetails>();
 
     public ClassLoaderDetails(UUID uuid, ClassLoaderSpec spec) {
         this.uuid = uuid;
