+++ platforms/core-runtime/launcher/src/main/java/org/gradle/tooling/internal/provider/ConnectionScopeServices.java
@@ -32,9 +32,9 @@ import org.gradle.launcher.daemon.configuration.DaemonParameters;
 import org.gradle.launcher.exec.BuildExecutor;
 import org.gradle.tooling.internal.adapter.ProtocolToModelAdapter;
 import org.gradle.tooling.internal.provider.serialization.ClassLoaderCache;
-import org.gradle.tooling.internal.provider.serialization.ClasspathInferer;
-import org.gradle.tooling.internal.provider.serialization.ClientSidePayloadClassLoaderFactory;
-import org.gradle.tooling.internal.provider.serialization.ClientSidePayloadClassLoaderRegistry;
+import org.gradle.internal.daemon.client.serialization.ClasspathInferer;
+import org.gradle.internal.daemon.client.serialization.ClientSidePayloadClassLoaderFactory;
+import org.gradle.internal.daemon.client.serialization.ClientSidePayloadClassLoaderRegistry;
 import org.gradle.tooling.internal.provider.serialization.DefaultPayloadClassLoaderRegistry;
 import org.gradle.tooling.internal.provider.serialization.ModelClassLoaderFactory;
 import org.gradle.tooling.internal.provider.serialization.PayloadSerializer;
